Refactored enemy classes and removed console logs
